(#112) Ensure custom cert on CCM Service
Prior to this commit the CCM service would generate a self-signed cert and use it even when a custom certificate was intended to be used.

With this commit the setup script now accepts a CertificateThumbprint parameter in which you can pass the
thumbprint of the certificate you wish to use.