(#223) Updates Chocolatey Install Script to match CLI Updates
The install script has been updated, and we should match that. This should bring it up to date with the changes made to the globally available version.

However, this eliminates some changes we have made to the script.

Consequently, this adds logic to the ClientSetup script to handle authenticated downloading of the specified  (or unspecified) version of Chocolatey from the client's Nexus instance. There are assumptions made about the repository software, here, and it may not work on another type of repository!