(#216) Cleanup jenkins setup to use new helpers
With the new helper functions in place, we can now with this new
code just call the helper functions directly, which really cleans
up and adds clarity to the code in the script for Jenkins setup.
The helper functions also give us greater control of the process
and should prevent edge cases of config file and README mismatches.

(squash)(#216)  Cleanup jenkins setup to use new helpers