(#158) Secure Jenkins installation with an SSL Certificate
Prior to this commit we were installing jenkins and configuring jobs, but doing
nothing with the web interface.

With this commit we use the same SSL certificate as other services and configure
Jenkins to run over HTTPS. This will allow folks to expose the interface more
broadly if they desire, and also provides an additional layer of security to the
overall installation of the guide.