(#86) Improve Script Output
As some of the changes have resulted in a lot less noise during the script(s) running, this adds a few more bits of output to show folk that progress is being made.

Conversely, this also removes a few of the block-outputs at the end of scripts that tell the user to login - as we don't want them to login until they've completed the guide (see #161).