(#262) Removal of ConsoleHost Check
The scripts all check if they're being run in something that isn't ConsoleHost. The warning there suggests this is because the scripts shouldn't be run in ISE, but after testing this they run correctly. It also allows for us to run the scripts in a remoting session, which is essential for the automated testing we're currently doing.

This change removes the requirement for running the quickstart scripts in a specific console host.