(#266) Updates Version Lookup in ClientSetup Script
The recent changes to Nexus cause the v2 lookup to no longer work correctly.

This commit changes the lookup to something that works on the post-3.71 available calls.

Though the v2-style assumed URL works currently, I have updated it to match the format returned by the index, as I can't speak for if Nexus will remove that.