(#285) Hardens ClientSetup Search Logic
The logic used in the client setup script to search for the latest version of Chocolatey available in the Repository URL had a minor flaw -

Using TrimEnd could result in repositories ending in any of the characters in "index" could trim too much.

This commit replaces that logic with regex, which hopefully will remove a problem.