(#291) Adds Nexus Configuration to Start-C4BNexus
Configuring the Nexus users and (optionally) SSL earlier in the Nexus script allows us to use the correct users and FQDNs immediately, and result in a more immediately available environment.

Now that we are always doing these actions (e.g. securing the repositories), this should also reduce a lot of confusion with what the SSL script does.