(#290) Removes usage of Web::GeneratePassword
This generator has no control over which characters are used and so was creating issues with our lazy replacements in the readme.

This removes all uses of it in favour of the method used in New-ServicePassword, which is more controlled.

Consequently, this commit also removes the New-CCMSalt function, which is now unused.