(#296) Fixes Self-Signed DatabaseUser Storage
When the initialize script was being run as a self-signed environment, it wasn't correctly generating and storing the DatabaseCredential.

This caused a prompt, later in the install flow, to enter a credential.

This change ensures that instead of checking for a ParameterSetName (which isn't set when no parameters are passed), it generates a credential if one is not provided already.